Africa Economic Analysis Report
Economic crises are getting common around the world and Africa is no exception. A rich continent yet faces numerous economic struggles.

To further understand the economic challenges, data was collected featuring the economic report of 13 countries in Africa from different year spans.

The report below shows how the 13 countries are affected by inflation, debt, cash crises, and banking crises.  It's obvious Zimbabwe and Angola seem to be at the receiving end of the crises.

The data was collected on kaggle and was visualized using power BI tool.