Tech Roles Salay Analysis
In an age were tech is taking over jobs around the world, tech roles under data science seems to be leading in the revolution of defining job roles.

This data is a collection of roles under data science outlining the salary and factors that determine the quality of job and the benefits attached.

From the visualization, it is obvious that:
Data science professionals earn more than other data roles. 
The more years of experience of an individual, the more the salary earned.