Financial Report
Financial report visualization of the sales and purchases in the Americas and Europe. 

From the data it can be observed that there are more sales in North America and Europe compared to South America. The graph also shows the peak sales months all year round which is observed to be in October. Finally, it seems government parastatals make more sales compared to the other segments in the continents.